pyproject.toml
[tool.poetry]
name = "meduzzen-backend"
version = "0.1.0"
description = ""
authors = ["FUZI <k.lipovok@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^5.1.2"
python-environ = "^0.4.54"
psycopg2 = "^2.9.10"
djangorestframework = "^3.15.2"
djoser = "^2.2.3"
django-cors-headers = "^4.6.0"
ruff = "^0.7.1"
pre-commit = "^4.0.1"
pytest = "^8.3.3"
django-extensions = "^3.2.3"
channels = "^4.2.0"
daphne = "^4.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
